deconcentrator/deconcentrator/requirements.txt
Django<3
djangorestframework
django-filter
django-redis
django-autoslug
celery
coreapi
Markdown
python-memcached
pyyaml
